+++ src/scripts/modules/common/Config.js
