+++ src/scripts/modules/content/ConfigManager.js
@@ -1,5 +1,5 @@
 // ConfigManager.js
-export class ConfigManager {
+class ConfigManager {
     static async loadConfig(defaultConfig) {
       return new Promise((resolve) => {
         chrome.storage.sync.get("jutsuExtensionConfig", (result) => {
@@ -17,4 +17,5 @@ export class ConfigManager {
       });
     }
   }
-  
+
+export default ConfigManager;
