+++ src/scripts/modules/content/NextSeriesManager.js
