+++ src/scripts/modules/content/ShortCatsManager.js
