+++ src/scripts/modules/content/SkipIntroManager.js
