+++ src/scripts/modules/content/VideoSpeedManager.js
